fix deps; refactor SCSS
- Remove unused styles and imports, streamline structure with @use for common and components
