+++ style/components/_index.scss
@@ -1,0 +1,9 @@
+@forward "about";
+@forward "achievements";
+@forward "brief";
+@forward "comments";
+@forward "history";
+@forward "info";
+@forward "lists";
+@forward "mobileActions";
+@forward "stats";
