+++ style/components/history.scss
@@ -1,65 +1,68 @@
-.c-history {
-    @extend %remove-margin;
-    width: 100% !important;
-    grid-column: 1 / 3;
+@use "../common" as *;
 
-    @media screen and (max-width: 767px) {
-        margin-bottom: $main-padding !important;
-        grid-column: 1;
-
-        .b-block_list {
-            display: grid;
-            grid-template-columns: minmax(100px, 1fr);
-        }
-    }
-
-    & > div:last-child,
-    .b-block_list {
-        @extend %block-style-hover;
-        @extend %border-radius-bottom;
-        @extend %subline-item;
-        display: grid;
-        grid-template-columns: repeat(auto-fit, minmax(100px, 1fr));
-        grid-gap: 1%;
+.profile-head {
+    .c-history {
+        @extend %remove-margin;
+        width: 100% !important;
+        grid-column: 1 / 3;
 
         @media screen and (max-width: 767px) {
-            display: grid;
-            grid-template-columns: minmax(100px, 1fr);
+            margin-bottom: $main-padding !important;
+            grid-column: 1;
+
+            .b-block_list {
+                display: grid;
+                grid-template-columns: minmax(100px, 1fr);
+            }
         }
 
-        .entry {
-            @extend %small-shadows-hover;
-            @extend %border-radius;
-            @extend %item-align;
-            @extend %remove-margin;
-            width: 100% !important;
-            transition: $history-entry-transition;
-            padding: 0 $small-padding !important;
-            justify-content: left !important;
-            border-top: 0 !important;
-            border: 0;
-            padding: 0;
+        &>div:last-child,
+        .b-block_list {
+            @extend %block-style-hover;
+            @extend %border-radius-bottom;
+            @extend %subline-item;
+            display: grid;
+            grid-template-columns: repeat(auto-fit, minmax(100px, 1fr));
+            grid-gap: 1%;
 
-            a {
-                min-width: 200px;
-                width: 100%;
+            @media screen and (max-width: 767px) {
+                display: grid;
+                grid-template-columns: minmax(100px, 1fr);
             }
 
-            .image-name {
+            .entry {
+                @extend %small-shadows-hover;
                 @extend %border-radius;
+                @extend %item-align;
+                @extend %remove-margin;
+                width: 100% !important;
+                transition: $history-entry-transition;
+                padding: 0 $small-padding !important;
+                justify-content: left !important;
+                border-top: 0 !important;
+                border: 0;
+                padding: 0;
 
-                img {
-                    @extend %border-radius;
-                    margin: calc($small-padding / 2) $small-padding
-                        calc($small-padding / 2) 0;
-                    max-height: 75px !important;
-                    max-width: 48px !important;
+                a {
+                    min-width: 200px;
+                    width: 100%;
                 }
-            }
 
-            &:hover {
-                text-decoration: none;
-                background: $entry-background;
+                .image-name {
+                    @extend %border-radius;
+
+                    img {
+                        @extend %border-radius;
+                        margin: calc($small-padding / 2) $small-padding calc($small-padding / 2) 0;
+                        max-height: 75px !important;
+                        max-width: 48px !important;
+                    }
+                }
+
+                &:hover {
+                    text-decoration: none;
+                    background: $entry-background;
+                }
             }
         }
     }
