+++ style/components/lists.scss
@@ -1,96 +1,102 @@
-.c-right {
-    width: 100% !important;
-    display: grid;
-    grid-auto-flow: row;
-    grid-gap: $main-padding;
+@use "../common" as *;
 
-    .cc-2a {
-        @extend %remove-sub-elements;
-        @extend %remove-margin;
-        display: grid;
-        grid-template-columns: 1fr 1fr;
-        grid-gap: $main-padding;
-        padding: 0;
-
-        > .c-column {
-            margin-right: 0;
-        }
-
-        .c-column {
-            width: 100%;
-        }
-    }
-
-    img {
-        @extend %border-radius;
-        @extend %remove-margin;
-        width: 100% !important;
-        max-width: none !important;
-        background: $entry-background;
-    }
-
-    %links {
-        a {
-            @extend %border-radius;
-            @extend %small-shadows;
-            @extend %small-shadows-hover;
-        }
-    }
-
-    .cc-friends {
-        @extend %block-style-hover;
-        @extend %border-radius-bottom;
-        @extend %links;
-        @extend %subline-item;
-        display: grid;
-        grid-template-columns: repeat(4, minmax(25px, 1fr));
-        grid-gap: 5px;
-        min-height: 150px;
-        height: 175px;
-
-        .b-user {
-            @extend %item-align;
-            padding: calc($small-padding / 2);
-        }
-
-        @extend %remove-sub-elements;
-    }
-
-    .b-clubs {
-        @extend %block-style-hover;
-        @extend %border-radius-bottom;
-        @extend %links;
-        @extend %remove-margin;
-        @extend %subline-item;
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-        grid-gap: 12px;
-        min-height: 150px;
-        height: 175px;
-
-        .b-club {
-            @extend %item-align;
-            @extend %remove-margin;
+.profile-content {
+    .cc-2.block {
+        .c-right {
             width: 100% !important;
+            display: grid;
+            grid-auto-flow: row;
+            grid-gap: $main-padding;
+
+            .cc-2a {
+                @extend %remove-sub-elements;
+                @extend %remove-margin;
+                display: grid;
+                grid-template-columns: 1fr 1fr;
+                grid-gap: $main-padding;
+                padding: 0;
+
+                >.c-column {
+                    margin-right: 0;
+                }
+
+                .c-column {
+                    width: 100%;
+                }
+            }
+
+            img {
+                @extend %border-radius;
+                @extend %remove-margin;
+                width: 100% !important;
+                max-width: none !important;
+                background: $entry-background;
+            }
+
+            %links {
+                a {
+                    @extend %border-radius;
+                    @extend %small-shadows;
+                    @extend %small-shadows-hover;
+                }
+            }
+
+            .cc-friends {
+                @extend %block-style-hover;
+                @extend %border-radius-bottom;
+                @extend %links;
+                @extend %subline-item;
+                display: grid;
+                grid-template-columns: repeat(4, minmax(25px, 1fr));
+                grid-gap: 5px;
+                min-height: 150px;
+                height: 175px;
+
+                .b-user {
+                    @extend %item-align;
+                    padding: calc($small-padding / 2);
+                }
+
+                @extend %remove-sub-elements;
+            }
+
+            .b-clubs {
+                @extend %block-style-hover;
+                @extend %border-radius-bottom;
+                @extend %links;
+                @extend %remove-margin;
+                @extend %subline-item;
+                display: grid;
+                grid-template-columns: repeat(2, 1fr);
+                grid-gap: 12px;
+                min-height: 150px;
+                height: 175px;
+
+                .b-club {
+                    @extend %item-align;
+                    @extend %remove-margin;
+                    width: 100% !important;
+                }
+
+                @extend %remove-sub-elements;
+            }
+
+            .cc-favourites {
+                @extend %block-style-hover;
+                @extend %border-radius-bottom;
+                @extend %links;
+                @extend %subline-item;
+                display: grid;
+                grid-gap: 10px;
+                grid-template-columns: repeat(auto-fit, minmax(70px, 1fr));
+
+                .b-catalog_entry {
+                    @extend %item-align;
+                    width: 100% !important;
+                }
+
+                @extend %remove-sub-elements;
+            }
         }
-
-        @extend %remove-sub-elements;
-    }
-
-    .cc-favourites {
-        @extend %block-style-hover;
-        @extend %border-radius-bottom;
-        @extend %links;
-        @extend %subline-item;
-        display: grid;
-        grid-gap: 10px;
-        grid-template-columns: repeat(auto-fit, minmax(70px, 1fr));
-
-        .b-catalog_entry {
-            @extend %item-align;
-            width: 100% !important;
-        }
-
-        @extend %remove-sub-elements;
     }
 }
